+++ test/runtests.jl
@@ -3,3 +3,4 @@ using OpeRe
 
 include("testtransportation.jl")
 include("testshortestpath.jl")
+include("testmaximumflow.jl")