+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.1.0 (Upcoming release)
 - Transportation problem definition
 - Linear programming solution of transportation problems 
-  
+- Linear programming solution of shortest path problem
+- Linear programming solution of maximum flow problem